requirements.txt
bcrypt==3.2.0
fastapi
fastapi[standard]
google-generativeai
jinja2
joblib
numpy==1.24.4
passlib
pydantic
python-dotenv
python-jose
python-multipart
protobuf==3.19.6
scikit-learn==1.3.2
scipy
sqlalchemy
tensorflow==2.10.0
uvicorn
